requirements-dev.txt
pytest==8.3.3
pytest-django==4.8.0
